requirements.txt
flask
flask_sqlalchemy
flask_wtf
python-dotenv
psycopg2-binary
email_validator
SQLAlchemy
flask-bcrypt
flask_login
psycopg2
Pillow
flask-mail
itsdangerous==1.1.0
markupsafe==2.0.1